api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
python-decouple==3.8
pydantic-settings==2.2.1
asyncpg==0.29.0
python-multipart==0.0.9
email-validator==2.1.1
python-jose==3.3.0
passlib==1.7.4